requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
crispy-bootstrap4==2024.1
dj-database-url==3.0.1
dj-email-url==1.0.6
Django==5.0.8
django-bootstrap-datepicker-plus==5.0.5
django-cache-url==3.4.5
django-ckeditor==6.7.1
django-ckeditor-5==0.2.18
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-filter==24.3
django-js-asset==2.2.0
django-taggit==6.0.0
djangorestframework==3.15.2
drf-yasg==1.21.10
environs==14.2.0
gunicorn==23.0.0
inflection==0.5.1
Markdown==3.6
marshmallow==4.0.0
packaging==25.0
pillow==10.4.0
psycopg2==2.9.10
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.2.0
whitenoise==6.9.0
